[index] Refactored index.js
