+++ es6/index.js
@@ -1,249 +1,194 @@
 "use strict";
 
 var SubContent = require("docxtemplater").SubContent;
-var FootNoteManager = require("./footNoteManager");
-var FootNoteReplacer = require("./footNoteReplacer");
+var XmlTemplater = require("docxtemplater").XmlTemplater;
+var FileTypeConfig = require("docxtemplater").FileTypeConfig;
+var DocUtils = require("./docUtils");
 
-class FootNoteModule {
-	constructor(options = {}) {
-		this.options = options;
-		// if (!(this.options.centered != null)) { this.options.centered = false; }
-		// if (!(this.options.getImage != null)) { throw new Error("You should pass getImage"); }
-		// if (!(this.options.getSize != null)) { throw new Error("You should pass getSize"); }
-		this.qrQueue = [];
-		this.imageNumber = 1;
+class FootnoteModule {
+	constructor(options) { 
+		this.options = options || {}; 
 	}
-	handleEvent(event, eventData) {
-		console.log("Handle event is called ");
-		console.log(event);
-		console.log(eventData);
-		console.log(this.manager);
+
+	
+	handleEvent(event) {
 		if (event === "rendering-file") {
-			this.renderingFileName = eventData;
 			var gen = this.manager.getInstance("gen");
-			this.footNoteManager = new FootNoteManager(gen.zip, this.renderingFileName);
-			console.log("About to load footNoteRels");
-			this.footNoteManager.loadFootNoteRels();
-		}
-		if (event === "rendered") {
-			if (this.qrQueue.length === 0) { return this.finished(); }
+			this.zip = gen.zip;
+			this.addStyles();
+			this.addStylesWithEffects();
+			this.updateSettings();
+			this.loadFootNoteRels();
+			this.addFootNoteContentType();
+			this.addFootNoteRels();
+			this.createFootNotesFile();
+			this.addEndNoteRels();
+			this.addEndNotesContentType()
+			this.createEndNotesFile();
+			return gen;
+		} else if (event === "rendered") {
+			return this.finished();
 		}
 	}
-	get(data) {
-		console.log("get(data) is called")
-		console.log(data)
-		if (data === "loopType") {
-			var templaterState = this.manager.getInstance("templaterState");
-			if (templaterState.textInsideTag[0] === "%") {
-				return "footnote";
-			}
+
+	loadFootNoteRels() {
+		var file = this.zip.files[`word/_rels/${this.endFileName}.xml.rels`] || this.zip.files["word/_rels/document.xml.rels"];
+		if (file === undefined) { return; }
+		var content = DocUtils.decodeUtf8(file.asText());
+		this.xmlDoc = DocUtils.str2xml(content);
+		// Get all Rids
+		var RidArray = [];
+		var iterable = this.xmlDoc.getElementsByTagName("Relationship");
+		for (var i = 0, tag; i < iterable.length - 1; i++) {
+			tag = iterable[i];
+			RidArray.push(parseInt(tag.getAttribute("Id").substr(3), 10));
 		}
+		this.maxRid = DocUtils.maxArray(RidArray);
+		return this;
+	}
+
+	addFootNoteContentType() {
+		var content = this.zip.files["[Content_Types].xml"].asText();
+		var xmlDoc = DocUtils.str2xml(content);
+		var addTag = true;
+		
+		var types = xmlDoc.getElementsByTagName("Types")[0];
+		var newTag = xmlDoc.createElement("Default");
+		newTag.namespaceURI = null;
+		newTag.setAttribute("ContentType", "application/vnd.openxmlformats-officedocument.wordprocessingml.footnotes+xml");
+		newTag.setAttribute("PartName", "/word/footnotes.xml");
+		types.appendChild(newTag);
+		this.updateFile("[Content_Types].xml", DocUtils.encodeUtf8(DocUtils.xml2Str(xmlDoc)));
+	}
+
+	addEndNotesContentType() {
+		var content = this.zip.files["[Content_Types].xml"].asText();
+		var xmlDoc = DocUtils.str2xml(content);
+		var addTag = true;
+		
+		var types = xmlDoc.getElementsByTagName("Types")[0];
+		var newTag = xmlDoc.createElement("Default");
+		newTag.namespaceURI = null;
+		newTag.setAttribute("ContentType", "application/vnd.openxmlformats-officedocument.wordprocessingml.endnotes+xml");
+		newTag.setAttribute("PartName", "/word/endnotes.xml");
+		types.appendChild(newTag);
+		this.updateFile("[Content_Types].xml", DocUtils.encodeUtf8(DocUtils.xml2Str(xmlDoc)));
+	}
+
+	addFootNoteRels() {
+		this.maxRid++;
+		var file = this.zip.files[`word/_rels/${this.endFileName}.xml.rels`] || this.zip.files["word/_rels/document.xml.rels"];
+		var content = DocUtils.decodeUtf8(file.asText());
+		var xmlDoc = DocUtils.str2xml(content);
+		var relationships = xmlDoc.getElementsByTagName("Relationships")[0];
+		var newTag = xmlDoc.createElement("Relationship");
+		newTag.namespaceURI = null;
+		newTag.setAttribute("Id", `rId${this.maxRid}`);
+		newTag.setAttribute("Type", "http://schemas.openxmlformats.org/officeDocument/2006/relationships/footnotes");
+		newTag.setAttribute("Target", 'footnotes.xml');
+		relationships.appendChild(newTag);
+		this.updateFile(file.name, DocUtils.encodeUtf8(DocUtils.xml2Str(xmlDoc)));
+	}
+
+	addEndNoteRels() {
+		this.maxRid++;
+		var file = this.zip.files[`word/_rels/${this.endFileName}.xml.rels`] || this.zip.files["word/_rels/document.xml.rels"];
+		var content = DocUtils.decodeUtf8(file.asText());
+		var xmlDoc = DocUtils.str2xml(content);
+		var relationships = xmlDoc.getElementsByTagName("Relationships")[0];
+		var newTag = xmlDoc.createElement("Relationship");
+		newTag.namespaceURI = null;
+		newTag.setAttribute("Id", `rId${this.maxRid}`);
+		newTag.setAttribute("Type", "http://schemas.openxmlformats.org/officeDocument/2006/relationships/endnotes");
+		newTag.setAttribute("Target", 'endnotes.xml');
+		relationships.appendChild(newTag);
+		this.updateFile(file.name, DocUtils.encodeUtf8(DocUtils.xml2Str(xmlDoc)));
+	}
+
+
+	createFootNotesFile() {
+		var prefix = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?><w:footnotes mc:Ignorable="w14 wp14" xmlns:m="http://schemas.openxmlformats.org/officeDocument/2006/math" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:mo="http://schemas.microsoft.com/office/mac/office/2008/main" xmlns:mv="urn:schemas-microsoft-com:mac:vml" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main" xmlns:w10="urn:schemas-microsoft-com:office:word" xmlns:w14="http://schemas.microsoft.com/office/word/2010/wordml" xmlns:wne="http://schemas.microsoft.com/office/word/2006/wordml" xmlns:wp="http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing" xmlns:wp14="http://schemas.microsoft.com/office/word/2010/wordprocessingDrawing" xmlns:wpc="http://schemas.microsoft.com/office/word/2010/wordprocessingCanvas" xmlns:wpg="http://schemas.microsoft.com/office/word/2010/wordprocessingGroup" xmlns:wpi="http://schemas.microsoft.com/office/word/2010/wordprocessingInk" xmlns:wps="http://schemas.microsoft.com/office/word/2010/wordprocessingShape"><w:footnote w:id="-1" w:type="separator"><w:pw:rsidR="002A7BE7" w:rsidRDefault="002A7BE7"><w:r><w:separator/></w:r></w:p></w:footnote><w:footnote w:id="0" w:type="continuationSeparator"><w:p w:rsidR="002A7BE7" w:rsidRDefault="002A7BE7"><w:r><w:continuationSeparator/></w:r></w:p></w:footnote>';
+		var content = this.addFootNotes();
+		var suffix = '</w:footnotes>';
+		var xmlString = prefix + content + suffix;
+
+		this.zip.file("word/footnotes.xml",xmlString, {});
+	}
+
+	updateFile(fileName, data, options = {}) {
+		this.zip.remove(fileName);
+		return this.zip.file(fileName, data, options);
+	}
+
+	addFootNotes() {
+		var output = "";
+		var footnotes = this.options.footnotes;
+		for (var counter = 0; counter < footnotes.length - 1; counter++) {
+			var referenceNumber = counter + 1
+			output += "<w:footnote w:id='" + referenceNumber + "'><w:p><w:pPr><w:pStyle w:val='FootnoteText'/></w:pPr><w:r><w:rPr><w:rStyle w:val='FootnoteReference'/></w:rPr><w:footnoteRef/></w:r><w:r><w:t xml:space='preserve'/></w:r><w:r><w:t>" + footnotes[counter] + "</w:t></w:r></w:p></w:footnote>"
+		}
+		return output;
+	}
+
+	createEndNotesFile() {
+		var xmlString = '<?xml version="1.0" encoding="UTF-8" standalone="yes"?><w:endnotes mc:Ignorable="w14 wp14" xmlns:m="http://schemas.openxmlformats.org/officeDocument/2006/math" xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" xmlns:mo="http://schemas.microsoft.com/office/mac/office/2008/main" xmlns:mv="urn:schemas-microsoft-com:mac:vml" xmlns:o="urn:schemas-microsoft-com:office:office" xmlns:r="http://schemas.openxmlformats.org/officeDocument/2006/relationships" xmlns:v="urn:schemas-microsoft-com:vml" xmlns:w="http://schemas.openxmlformats.org/wordprocessingml/2006/main" xmlns:w10="urn:schemas-microsoft-com:office:word" xmlns:w14="http://schemas.microsoft.com/office/word/2010/wordml" xmlns:wne="http://schemas.microsoft.com/office/word/2006/wordml" xmlns:wp="http://schemas.openxmlformats.org/drawingml/2006/wordprocessingDrawing" xmlns:wp14="http://schemas.microsoft.com/office/word/2010/wordprocessingDrawing" xmlns:wpc="http://schemas.microsoft.com/office/word/2010/wordprocessingCanvas" xmlns:wpg="http://schemas.microsoft.com/office/word/2010/wordprocessingGroup" xmlns:wpi="http://schemas.microsoft.com/office/word/2010/wordprocessingInk" xmlns:wps="http://schemas.microsoft.com/office/word/2010/wordprocessingShape"><w:endnote w:id="-1" w:type="separator"><w:p><w:r><w:separator/></w:r></w:p></w:endnote><w:endnote w:id="0" w:type="continuationSeparator"><w:p><w:r><w:continuationSeparator/></w:r></w:p></w:endnote></w:endnotes>'
+		this.zip.file("word/endnotes.xml",xmlString, {});
+	}
+
+	addStyles() {
+		var file = this.zip.files["word/styles.xml"];
+		var xmlString = DocUtils.decodeUtf8(file.asText());
+		xmlString = xmlString.replace("</w:styles>", "<w:style w:styleId='FootnoteText' w:type='paragraph'><w:name w:val='footnote text'/><w:basedOn w:val='Normal'/><w:link w:val='FootnoteTextChar'/><w:uiPriority w:val='99'/><w:unhideWhenUsed/></w:style><w:style w:customStyle='1' w:styleId='FootnoteTextChar' w:type='character'><w:name w:val='Footnote Text Char'/><w:basedOn w:val='DefaultParagraphFont'/><w:link w:val='FootnoteText'/><w:uiPriority w:val='99'/></w:style><w:style w:styleId='FootnoteReference' w:type='character'><w:name w:val='footnote reference'/><w:basedOn w:val='DefaultParagraphFont'/><w:uiPriority w:val='99'/><w:unhideWhenUsed/><w:rPr><w:vertAlign w:val='superscript'/></w:rPr></w:style></w:styles>")
+		this.updateFile("word/styles.xml",xmlString);
+
+	}
+
+	addStylesWithEffects() {
+		var file = this.zip.files["word/stylesWithEffects.xml"];
+		var xmlString = DocUtils.decodeUtf8(file.asText());
+		xmlString = xmlString.replace("</w:styles>", "<w:style w:styleId='FootnoteText' w:type='paragraph'><w:name w:val='footnote text'/><w:basedOn w:val='Normal'/><w:link w:val='FootnoteTextChar'/><w:uiPriority w:val='99'/><w:unhideWhenUsed/></w:style><w:style w:customStyle='1' w:styleId='FootnoteTextChar' w:type='character'><w:name w:val='Footnote Text Char'/><w:basedOn w:val='DefaultParagraphFont'/><w:link w:val='FootnoteText'/><w:uiPriority w:val='99'/></w:style><w:style w:styleId='FootnoteReference' w:type='character'><w:name w:val='footnote reference'/><w:basedOn w:val='DefaultParagraphFont'/><w:uiPriority w:val='99'/><w:unhideWhenUsed/><w:rPr><w:vertAlign w:val='superscript'/></w:rPr></w:style></w:styles>")
+		this.updateFile("word/stylesWithEffects.xml",xmlString);
+
+	}
+
+	updateSettings() {
+		var file = this.zip.files["word/settings.xml"];
+		var xmlString = DocUtils.decodeUtf8(file.asText());
+		xmlString = xmlString.replace("</w:settings>","<w:footnotePr><w:footnote w:id='-1'/><w:footnote w:id='0'/></w:footnotePr><w:endnotePr><w:endnote w:id='-1'/><w:endnote w:id='0'/></w:endnotePr></w:settings>");
+		// var xmlDoc = DocUtils.str2xml(xmlString);
+		// var settings = xmlDoc.getElementsByTagName("w:settings")[0];
+		// var rsids = xmlDoc.getElementsByTagName("w:rsids")[0];
+		// var newTag = xmlDoc.createElement("w:rsid");
+		// newTag.namespaceURI = null;
+		// newTag.setAttribute("w:val", "002A7BE7");
+		// rsids.appendChild(newTag);
+		// this.updateFile("word/settings.xml",DocUtils.encodeUtf8(DocUtils.xml2Str(xmlDoc)));
+		this.updateFile("word/settings.xml",xmlString);
+	}
+ 
+	get(data) {
 		return null;
 	}
-	getNextImageName() {
-		var name = `image_generated_${this.imageNumber}.png`;
-		this.imageNumber++;
-		return name;
-	}
-	replaceBy(text, outsideElement) {
-		var xmlTemplater = this.manager.getInstance("xmlTemplater");
-		var templaterState = this.manager.getInstance("templaterState");
-		var subContent = new SubContent(xmlTemplater.content);
-		subContent = subContent.getInnerTag(templaterState);
-		subContent = subContent.getOuterXml(outsideElement);
-		return xmlTemplater.replaceXml(subContent, text);
-	}
-	convertPixelsToEmus(pixel) {
-		return Math.round(pixel * 9525);
-	}
-	replaceTag() {
-		console.log("replaceTag")
-		var scopeManager = this.manager.getInstance("scopeManager");
-		var templaterState = this.manager.getInstance("templaterState");
-		var xmlTemplater = this.manager.getInstance("xmlTemplater");
-		var tagXml = xmlTemplater.fileTypeConfig.tagsXmlArray[0];
 
-		var tag = templaterState.textInsideTag.substr(1);
-		var tagValue = scopeManager.getValue(tag);
-
-		if (tagValue == null) {
-			return this.replaceBy(startEnd, tagXml);
-		}
-
-		var tagXmlParagraph = tagXml.substr(0, 1) + ":p";
-
-		var startEnd = `<${tagXml}></${tagXml}>`;
-		var footNoteRels = this.footNoteManager.loadFootNoteRels();
-		if (!footNoteRels) {
-			return;
-		}
-		footNoteRels.addFootNote();
-		var imgBuffer;
-
-		try {
-			imgBuffer = this.options.getImage(tagValue, tag);
-		}
-		catch (e) {
-			return this.replaceBy(startEnd, tagXml);
-		}
-		
-		var rId = footNoteRels.addImageRels(this.getNextImageName(), imgBuffer);
-		var sizePixel = this.options.getSize(imgBuffer, tagValue, tag);
-		var size = [this.convertPixelsToEmus(sizePixel[0]), this.convertPixelsToEmus(sizePixel[1])];
-		var newText = this.options.centered ? this.getImageXmlCentered(rId, size) : this.getImageXml(rId, size);
-		var outsideElement = this.options.centered ? tagXmlParagraph : tagXml;
-		return this.replaceBy(newText, outsideElement);
+	handle(type, data) {
+		return null;
 	}
-	replaceQr() {
-		var xmlTemplater = this.manager.getInstance("xmlTemplater");
-		var imR = new FootNoteReplacer(xmlTemplater, this.footNoteManager);
-		imR.getDataFromString = (result, cb) => {
-			if ((this.options.getImageAsync != null)) {
-				return this.options.getImageAsync(result, cb);
-			}
-			return cb(null, this.options.getImage(result));
-		};
-		imR.pushQrQueue = (num) => {
-			return this.qrQueue.push(num);
-		};
-		imR.popQrQueue = (num) => {
-			var found = this.qrQueue.indexOf(num);
-			if (found !== -1) {
-				this.qrQueue.splice(found, 1);
-			}
-			else {
-				this.on("error", new Error(`qrqueue ${num} is not in qrqueue`));
-			}
-			if (this.qrQueue.length === 0) { return this.finished(); }
-		};
-		var num = parseInt(Math.random() * 10000, 10);
-		imR.pushQrQueue("rendered-" + num);
-		try {
-			imR.findImages().replaceImages();
-		}
-		catch (e) {
-			this.on("error", e);
-		}
-		var f = () => imR.popQrQueue("rendered-" + num);
-		return setTimeout(f, 1);
-	}
+
 	finished() {}
+
 	on(event, data) {
 		if (event === "error") {
 			throw data;
 		}
 	}
-	handle(type, data) {
-		if (type === "replaceTag" && data === "footnote") {
-			this.replaceTag();
-		}
-		if (type === "xmlRendered" && this.options.qrCode) {
-			this.replaceQr();
-		}
-		return null;
+
+	replaceBy(text, outsideElement) {
+		return text;
 	}
-	getImageXml(rId, size) {
-		return `<w:drawing>
-  <wp:inline distT="0" distB="0" distL="0" distR="0">
-    <wp:extent cx="${size[0]}" cy="${size[1]}"/>
-    <wp:effectExtent l="0" t="0" r="0" b="0"/>
-    <wp:docPr id="2" name="Image 2" descr="image"/>
-    <wp:cNvGraphicFramePr>
-      <a:graphicFrameLocks xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main" noChangeAspect="1"/>
-    </wp:cNvGraphicFramePr>
-    <a:graphic xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-      <a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/picture">
-        <pic:pic xmlns:pic="http://schemas.openxmlformats.org/drawingml/2006/picture">
-          <pic:nvPicPr>
-            <pic:cNvPr id="0" name="Picture 1" descr="image"/>
-            <pic:cNvPicPr>
-              <a:picLocks noChangeAspect="1" noChangeArrowheads="1"/>
-            </pic:cNvPicPr>
-          </pic:nvPicPr>
-          <pic:blipFill>
-            <a:blip r:embed="rId${rId}">
-              <a:extLst>
-                <a:ext uri="{28A0092B-C50C-407E-A947-70E740481C1C}">
-                  <a14:useLocalDpi xmlns:a14="http://schemas.microsoft.com/office/drawing/2010/main" val="0"/>
-                </a:ext>
-              </a:extLst>
-            </a:blip>
-            <a:srcRect/>
-            <a:stretch>
-              <a:fillRect/>
-            </a:stretch>
-          </pic:blipFill>
-          <pic:spPr bwMode="auto">
-            <a:xfrm>
-              <a:off x="0" y="0"/>
-              <a:ext cx="${size[0]}" cy="${size[1]}"/>
-            </a:xfrm>
-            <a:prstGeom prst="rect">
-              <a:avLst/>
-            </a:prstGeom>
-            <a:noFill/>
-            <a:ln>
-              <a:noFill/>
-            </a:ln>
-          </pic:spPr>
-        </pic:pic>
-      </a:graphicData>
-    </a:graphic>
-  </wp:inline>
-</w:drawing>
-		`;
-	}
-	getImageXmlCentered(rId, size) {
-		return `		<w:p>
-		  <w:pPr>
-			<w:jc w:val="center"/>
-		  </w:pPr>
-		  <w:r>
-			<w:rPr/>
-			<w:drawing>
-			  <wp:inline distT="0" distB="0" distL="0" distR="0">
-				<wp:extent cx="${size[0]}" cy="${size[1]}"/>
-				<wp:docPr id="0" name="Picture" descr=""/>
-				<a:graphic xmlns:a="http://schemas.openxmlformats.org/drawingml/2006/main">
-				  <a:graphicData uri="http://schemas.openxmlformats.org/drawingml/2006/picture">
-					<pic:pic xmlns:pic="http://schemas.openxmlformats.org/drawingml/2006/picture">
-					  <pic:nvPicPr>
-						<pic:cNvPr id="0" name="Picture" descr=""/>
-						<pic:cNvPicPr>
-						  <a:picLocks noChangeAspect="1" noChangeArrowheads="1"/>
-						</pic:cNvPicPr>
-					  </pic:nvPicPr>
-					  <pic:blipFill>
-						<a:blip r:embed="rId${rId}"/>
-						<a:stretch>
-						  <a:fillRect/>
-						</a:stretch>
-					  </pic:blipFill>
-					  <pic:spPr bwMode="auto">
-						<a:xfrm>
-						  <a:off x="0" y="0"/>
-						  <a:ext cx="${size[0]}" cy="${size[1]}"/>
-						</a:xfrm>
-						<a:prstGeom prst="rect">
-						  <a:avLst/>
-						</a:prstGeom>
-						<a:noFill/>
-						<a:ln w="9525">
-						  <a:noFill/>
-						  <a:miter lim="800000"/>
-						  <a:headEnd/>
-						  <a:tailEnd/>
-						</a:ln>
-					  </pic:spPr>
-					</pic:pic>
-				  </a:graphicData>
-				</a:graphic>
-			  </wp:inline>
-			</w:drawing>
-		  </w:r>
-		</w:p>
-		`;
+	
+	replaceTag() {
+		return "";
 	}
 }
 
-module.exports = FootNoteModule;
+module.exports = FootnoteModule;
